Reverted rewardAmount back to float to maintain consistency with original plan

+++ Code/project/src/main/java/com/mycompany/app/classes/Reward.java
@@ -3,20 +3,18 @@ package com.mycompany.app.classes;
 public class Reward extends StaticEntity 
 {
 
-    //I changed the type for rewardAmount from float to double
-    private double rewardAmount;
+    private float rewardAmount;
     private int rewardsToCollect;
     public RewardType rewardType;
 
-
-    public double getRewardAmount()
+    public float getRewardAmount()
     {
         return rewardAmount;
     }
 
     public void setRewardAmount()
     {
-        this.rewardAmount = 0.5;
+        this.rewardAmount = 0.5f;
     }
 
     public int getRewardsToCollect()
